+++ modules/fx_ai_chat/lib/src/model/model.dart
@@ -1,0 +1,4 @@
+export 'ai_chat_input.dart';
+export 'ai_chat_output.dart';
+export 'ai_choice.dart';
+export 'message.dart';