+++ modules/fx_platform_adapter/lib/src/data/global.dart
@@ -1,9 +1,7 @@
+import 'package:fx_env/fx_env.dart';
 import 'package:window_manager/window_manager.dart' as manager;
 import 'package:window_manager/window_manager.dart';
 
-import '../app_env/app_env.dart';
-
-
 WindowManager windowManager = manager.windowManager;
 
-AppEnv kAppEnv = AppEnv();
+AppEnv kAppEnv = kApp;