+++ modules/fx_string/lib/fx_string.dart
@@ -1,0 +1,2 @@
+library;
+export 'src/string.dart';