+++ modules/fx_trace/lib/fx_trace.dart
@@ -2,3 +2,6 @@ library fx_trace;
 
 export 'src/mixin.dart';
 export 'src/model/model.dart';
+export 'src/emitter/fx_event.dart';
+export 'src/emitter/fx_emitter.dart';
+export 'src/emitter/event_ticker_mixin.dart';