+++ modules/fx_trace/lib/src/emitter/fx_event.dart
@@ -1,0 +1,3 @@
+class FxEvent {
+  const FxEvent();
+}