+++ modules/fx_trace/lib/src/model/trace/trace.dart
@@ -1,6 +1,8 @@
+import 'package:fx_trace/fx_trace.dart';
+
 import '../model.dart';
 
-mixin Trace {
+mixin Trace implements Exception {
   Code? get code;
 
   String? get message;
@@ -8,4 +10,6 @@ mixin Trace {
   StackTrace? get stack;
 
   Object? get error;
+
+  void emit() => FxTrace().emit(this);
 }