+++ modules/fx_trace/CHANGELOG.md
@@ -7,3 +7,11 @@
 
 * add [TipLevel] in TipTrace
 * download dart sdk version: min 3.4.0
+
+## 0.0.3
+
+* add [FxEvent]/[FxEmitter]/[FxEmitterMixin]
+
+## 0.0.4
+
+* add [Trace#emit]/[LogTrace]